% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{fit.yapa}
\alias{fit.yapa}
\title{Fit the model}
\usage{
fit.yapa(m, ...)
}
\arguments{
\item{m}{list of inputs to the model}

\item{...}{Additional arguments}
}
\description{
This sets m$params to contain the fitted model parameters.
}
