% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grass_gss}
\alias{grass_gss}
\title{GSS results for the "grass" question.}
\format{
A data frame with 292 rows and 5 variables:
\describe{
\item{year}{year}
\item{n}{sample size}
\item{legal}{total respondents who said legal}
\item{illegal}{total respondents who said illegal}
}
}
\source{
\url{www.https://gss.norc.org/}
}
\usage{
grass_gss
}
\description{
A dataset containing the responses to the General Social Survey for the question "Should marijuana be made legal?"
}
\keyword{datasets}
