% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{make_output_dataframe}
\alias{make_output_dataframe}
\title{Helper function for formatting output}
\usage{
make_output_dataframe(var, nrow, ncol, y, dates)
}
\arguments{
\item{var}{the outcome variable.}

\item{nrow}{the number of observations.}

\item{ncol}{the number of options.}

\item{y}{the poll counts matrix.}

\item{dates}{the dates variable.}
}
\value{
dataframe of tidied outputs
}
\description{
Helper function for formatting output
}
\keyword{internal}
