% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obama_mccain_polls}
\alias{obama_mccain_polls}
\title{Polls for the 2008 presidential race between Obama and McCain.}
\format{
A data frame with 292 rows and 5 variables:
\describe{
\item{pollster}{pollster}
\item{poll_end}{the end date}
\item{n}{sample size}
\item{\code{McCain (R)}}{total respondents who said McCain}
\item{\code{Obama (D)}}{total respondents who said Obama}
}
}
\source{
\url{www.realclearpolitics.com}
}
\usage{
obama_mccain_polls
}
\description{
A dataset containing the pollster, date, sample size, and results for Obama and McCain in 292 polls.
}
\keyword{datasets}
