% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{plot_yapa}
\alias{plot_yapa}
\title{Function to plot yapa trend, modeled averages, and raw data.}
\usage{
plot_yapa(
  yapafit,
  size = 1,
  se = TRUE,
  poll_alpha = 0.2,
  model_alpha = 1,
  se_alpha = 0.1,
  ...
)
}
\arguments{
\item{yapafit}{an object of class yapafit}
}
\value{
a ggplot
}
\description{
Function to plot yapa trend, modeled averages, and raw data.
}
\keyword{internal}
