% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollagg-package.R
\docType{package}
\name{pollagg-package}
\alias{pollagg-package}
\alias{pollagg}
\title{The 'pollagg' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. http://mc-stan.org
}
