% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{print.yapafit}
\alias{print.yapafit}
\title{print a yapa model}
\usage{
\method{print}{yapafit}(x, ...)
}
\arguments{
\item{x}{a yapafit model.}

\item{...}{additional arguments.}
}
\description{
print a yapa model
}
