% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{tidy_polls}
\alias{tidy_polls}
\title{Helper function for formatting output}
\usage{
tidy_polls(y, n, dates)
}
\arguments{
\item{y}{the poll counts matrix.}

\item{n}{the poll sample sizes.}

\item{dates}{the dates variable.}
}
\value{
dataframe of tidied outputs
}
\description{
Helper function for formatting output
}
\keyword{internal}
