% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{validate}
\alias{validate}
\title{Validates input to yapa model}
\usage{
validate(m)
}
\arguments{
\item{m}{list of inputs.}
}
\value{
list of inputs if they have passed validation
}
\description{
Validates input to yapa model
}
\keyword{internal}
