% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{summary.yapafit}
\alias{summary.yapafit}
\title{summarize a yapa model}
\usage{
\method{summary}{yapafit}(object, ...)
}
\arguments{
\item{object}{a yapafit model.}

\item{...}{additional arguments.}
}
\description{
summarize a yapa model
}
