% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapa.R
\name{yapa}
\alias{yapa}
\title{fit a yapa model}
\usage{
yapa(y, n, poll_dates, dates = NULL)
}
\arguments{
\item{y}{a matrix of poll respondent counts.}

\item{n}{a vector of poll sample sizes.}

\item{dates}{a vector of poll dates.}
}
\description{
fit a yapa model
}
